apps/api/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "ai-rakuraku-api"
version = "0.1.0"
description = "FastAPI backend for AI Rakuraku Ordermade"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "supabase==2.4.1",
    "redis==5.0.1",
    "stripe==7.4.0",
    "python-multipart==0.0.6",
    "trafilatura==1.6.3",
    "pymupdf==1.23.8",
    "pydantic-settings==2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.1",
]
